+++ src/main/java/br/com/vitavault/domain/MovimentacaoEstoque.java
@@ -17,7 +17,7 @@ public class MovimentacaoEstoque {
     private LocalDate dataMovimentacao;
     private EnumTipoMovimentacao tipoMovimentacao;
     private Long quantidade;
-    private Estoque estoque;
+    // private Estoque estoque;
 
     /*
         Autora: Ariadne Cavilha Jorge
@@ -33,34 +33,39 @@ public class MovimentacaoEstoque {
     }
 
     public MovimentacaoEstoque(Produto produto, Funcionario funcionario, Papel alcada, LocalDate dataMovimentacao, EnumTipoMovimentacao tipoMovimentacao, Long quantidade) {
-        this(new ItemEstoque(UUID.randomUUID(), produto, LocalDate.now(), quantidade, tipoMovimentacao), funcionario, alcada, dataMovimentacao, tipoMovimentacao, quantidade);
+        this(new ItemEstoque(UUID.randomUUID(), produto, LocalDate.now(), 0L, tipoMovimentacao), funcionario, alcada, dataMovimentacao, tipoMovimentacao, quantidade);
     }
 
     /*
         Autora: Ariadne Cavilha Jorge
     */
     public void adicionaMovimentacao(EnumTipoMovimentacao tipoMovimentacao, Produto produto, ItemEstoque itemEstoque, Long quantidade) {
-        Estoque estoque = new Estoque();
+        //Estoque estoque = new Estoque();
 
         if (tipoMovimentacao == EnumTipoMovimentacao.ENTRADA) {
-            movimentarEntrada(produto, itemEstoque, quantidade, estoque);
+            //  movimentarEntrada(produto, itemEstoque, quantidade, estoque);
         } else {
             // implementar o método movimentarSaida
         }
     }
 
-    private void movimentarEntrada(Produto produto, ItemEstoque itemEstoque, Long quantidade, Estoque estoque) {
-        estoque.vincularEstoqueAoItem(itemEstoque, estoque);
-        setEstoque(estoque);
-        //setarQuantidade(itemEstoque.getQuantidade(), quantidade, itemEstoque);
-        itemEstoque.setQuantidade(quantidade);
-        estoque.adicionarProduto(produto.getId(), itemEstoque);
-        System.out.println("Id do produto no item estoque " + produto.getId());
+    public void movimentarEntrada(ItemEstoque itemEstoque, Long quantidade, Estoque estoque) {
+        itemEstoque = estoque.vincularEstoqueAoItem(itemEstoque, estoque);
+
+        vincularMovimentacaoAoItem(itemEstoque);
+        setarQuantidade(itemEstoque, quantidade);
+
+
+        System.out.println("Id do item estoque " + itemEstoque.getId());
         System.out.println("Saldo atual: " + itemEstoque.getQuantidade());
     }
 
-    private void setarQuantidade(Long quantidadeEstoque, Long quantidadeAumentar, ItemEstoque itemEstoque) {
-        itemEstoque.setQuantidade(quantidadeEstoque + quantidadeAumentar);
+    private void vincularMovimentacaoAoItem(ItemEstoque itemEstoque) {
+        itemEstoque.getMovimentacao().add(this);
+    }
+
+    private void setarQuantidade(ItemEstoque itemEstoque, Long quantidadeAumentar) {
+        itemEstoque.setQuantidade(itemEstoque.getQuantidade() + quantidadeAumentar);
     }
 
     public Long getQuantidade() {
@@ -75,11 +80,11 @@ public class MovimentacaoEstoque {
         return item;
     }
 
-    public void setEstoque(Estoque estoque) {
-        this.estoque = estoque;
-    }
+//    public void setEstoque(Estoque estoque) {
+//        this.estoque = estoque;
+//    }
 
-    public void listarProdutosEstoque() {
-        estoque.listarProdutosEstoque();
-    }
+//    public void listarProdutosEstoque() {
+//        estoque.listarProdutosEstoque();
+//    }
 }